Refactor database into module
