+++ fast_track/database/__init__.py
@@ -1,0 +1,3 @@
+""" Initializes the db module. """
+
+from .database import Database
