+++ fast_track/database/schemas/base.py
