+++ fast_track/database/schemas/detection.py
