+++ run.py
@@ -7,6 +7,7 @@ from dotenv import load_dotenv
 from fast_track import Pipeline
 from fast_track.detectors import YOLONAS
 from fast_track.trackers import BYTETracker
+from fast_track.database import Database
 
 load_dotenv()
 
@@ -16,7 +17,14 @@ with open('config/coco.yml', 'r') as f:
 
 camera = cv2.VideoCapture(config['data_path'])
 detector = YOLONAS(**config['detector'], names=config['names'], image_shape=(camera.get(3), camera.get(4)))
-tracker = BYTETracker(**config['tracker'], names=config['names'], **config["db"])
+tracker = BYTETracker(**config['tracker'], names=config['names'])
+database = Database(**config["db"], class_names=config['names'])
 
-with Pipeline(camera=camera, detector=detector, tracker=tracker, outfile=config['outfile']) as p:
-    p.run()
+with Pipeline(
+    camera=camera,
+    detector=detector,
+    tracker=tracker,
+    database=database,
+    outfile=config['outfile']
+) as pipeline:
+    pipeline.run()
